store/requirements.txt
asgiref==3.4.1
black==21.12b0
click==8.0.3
Django==3.2
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
